Safeconvert: add arm specific code for int to uint32
On arm, there is no need to check if an int is bigger than uint32 as it
will never be the case. Additionally, leaving the check would cause
building on arm to fail.
